backend/requirements.txt
Django==3.0.5
django-3-jet==1.0.8
django-filter==2.2.0
django-mptt==0.11.0
django-nested-admin==3.2.4
djangorestframework==3.11.0
djangorestframework-filters==1.0.0.dev0
django-cors-headers==3.2.1
Pillow==7.1.1
djangorestframework-camel-case==1.1.2
django-admin-numeric-filter==0.1.6
# wemake-python-styleguide==0.14.0
